backend/requirements.txt
Django==4.2.10
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-allauth==0.60.1
django-anymail[sendgrid]==10.3
stripe==7.14.0
paypalrestsdk==1.13.1
celery==5.3.6
redis==5.0.1
django-cors-headers==4.3.1
sentry-sdk==1.39.1
weasyprint==61.1
python-dotenv==1.0.1
gunicorn==21.2.0
django-redis==5.4.0
Pillow==10.0.1
django-filter==23.5
drf-spectacular==0.26.5
whitenoise==6.6.0